# Simplify jpeg encoder and fix for Windows ARM