--- a/src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
@@ -376,16 +376,16 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components
         /// <param name="block">Source block</param>
         /// <param name="dest">Destination block</param>
         /// <param name="qt">The quantization table</param>
-        /// <param name="unzigPtr">Pointer to elements of <see cref="ZigZag"/></param>
+        /// <param name="unZig">The 8x8 Unzig block.</param>
         public static unsafe void Quantize(
             ref Block8x8F block,
             ref Block8x8F dest,
             ref Block8x8F qt,
-            byte* unzigPtr)
+            ref ZigZag unZig)
         {
             for (int zig = 0; zig < Size; zig++)
             {
-                dest[zig] = block[unzigPtr[zig]];
+                dest[zig] = block[unZig[zig]];
             }
 
             DivideRoundAll(ref dest, ref qt);
@@ -396,12 +396,12 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components
         /// </summary>
         /// <param name="destination">The destination block.</param>
         /// <param name="source">The source block.</param>
-        public static unsafe void Scale16X16To8X8(ref Block8x8F destination,  Block8x8F* source)
+        public static unsafe void Scale16X16To8X8(ref Block8x8F destination, ReadOnlySpan<Block8x8F> source)
         {
             for (int i = 0; i < 4; i++)
             {
                 int dstOff = ((i & 2) << 4) | ((i & 1) << 2);
-                float* iSource = (float*)(source + i);
+                Block8x8F iSource = source[i];
 
                 for (int y = 0; y < 4; y++)
                 {