--- a/src/ImageSharp/Formats/Jpeg/Components/Encoder/BlockQuad.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Encoder/BlockQuad.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors.
+// Copyright (c) Six Labors.
 // Licensed under the Apache License, Version 2.0.
 
 namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Encoder