--- a/src/ImageSharp/Formats/Jpeg/Components/Encoder/YCbCrForwardConverter{TPixel}.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Encoder/YCbCrForwardConverter{TPixel}.cs
@@ -62,9 +62,9 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Encoder
             Span<Rgb24> rgbSpan = this.rgbBlock.AsSpanUnsafe();
             PixelOperations<TPixel>.Instance.ToRgb24(frame.GetConfiguration(), this.pixelBlock.AsSpanUnsafe(), rgbSpan);
 
-            ref float yBlockStart = ref Unsafe.As<Block8x8F, float>(ref this.Y);
-            ref float cbBlockStart = ref Unsafe.As<Block8x8F, float>(ref this.Cb);
-            ref float crBlockStart = ref Unsafe.As<Block8x8F, float>(ref this.Cr);
+            ref Block8x8F yBlock = ref this.Y;
+            ref Block8x8F cbBlock = ref this.Cb;
+            ref Block8x8F crBlock = ref this.Cr;
             ref Rgb24 rgbStart = ref rgbSpan[0];
 
             for (int i = 0; i < 64; i++)
@@ -75,9 +75,10 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Encoder
                     c.R,
                     c.G,
                     c.B,
-                    ref Unsafe.Add(ref yBlockStart, i),
-                    ref Unsafe.Add(ref cbBlockStart, i),
-                    ref Unsafe.Add(ref crBlockStart, i));
+                    ref yBlock,
+                    ref cbBlock,
+                    ref crBlock,
+                    i);
             }
         }
     }