--- a/src/ImageSharp/PixelFormats/PixelImplementations/Rgb24.cs
+++ b/src/ImageSharp/PixelFormats/PixelImplementations/Rgb24.cs
@@ -204,7 +204,6 @@ namespace SixLabors.ImageSharp.PixelFormats
         public void FromRgba32(Rgba32 source)
         {
 #if NETSTANDARD2_0
-
             // See https://github.com/SixLabors/ImageSharp/issues/1275
             this.R = source.R;
             this.G = source.G;