--- a/tests/ImageSharp.Benchmarks/Codecs/Jpeg/EncodeJpeg.cs
+++ b/tests/ImageSharp.Benchmarks/Codecs/Jpeg/EncodeJpeg.cs
@@ -63,15 +63,14 @@ namespace SixLabors.ImageSharp.Benchmarks.Codecs.Jpeg
 
 /*
 BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.959 (1909/November2018Update/19H2)
-Intel Core i7-8650U CPU 1.90GHz(Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
-.NET Core SDK = 3.1.302
-
- [Host]     : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT
+Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
+.NET Core SDK=3.1.302
+  [Host]     : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT
   DefaultJob : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT
 
 
-|                Method |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD |
-|---------------------- |---------:|----------:|----------:|---------:|------:|--------:|
-| 'System.Drawing Jpeg' | 4.473 ms | 0.0847 ms | 0.2012 ms | 4.408 ms |  1.00 |    0.00 |
-|     'ImageSharp Jpeg' | 5.409 ms | 0.0379 ms | 0.0316 ms | 5.412 ms |  1.19 |    0.05 |
+|                Method |     Mean |     Error |    StdDev | Ratio | RatioSD |
+|---------------------- |---------:|----------:|----------:|------:|--------:|
+| 'System.Drawing Jpeg' | 4.297 ms | 0.0244 ms | 0.0228 ms |  1.00 |    0.00 |
+|     'ImageSharp Jpeg' | 5.286 ms | 0.1034 ms | 0.0967 ms |  1.23 |    0.02 |
 */