### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Fixes #1275 

- [x] Use safe cast to span where possible
- [x] Simplify pointer code to use known safe paths used during decoding.
- [x] Limit `Unsafe` API usage
- [x] Use simplified code to debug issues
- [x] Prove it works!

<!-- Thanks for contributing to ImageSharp! -->