@JimBobSquarePants 2020-08-06
>Can we do a before/after benchmark comparison for EncodeJpeg, with an image different than Car.bmp?

I suggest re-encoding jpeg420exif.jpg, but config can be Config.ShortCore31 to get it done faster.

@antonfirsov I'm not all that fussed if we lose a percentile or two here. We need to create a hardware accelerated version of our transforms anyway.