@antonfirsov 2020-08-06
> if we lose a percentile or two here

I'm afraid it might be more, that's why I suggest an extensive benchmark run. Concerned about `Block8x8F` indexer performance, since it's used very frequently now, and it's code doesn't seem optimal to me.