@JimBobSquarePants 2020-08-06
>I'm afraid it might be more, that's why I suggest an extensive benchmark run. Concerned about Block8x8F indexer performance, since it's used very frequently now, and it's code doesn't seem optimal to me.

We use it in a few places in the decoder, hence why I chose to reuse it as a known path. 

Anyway.... Here's the before and after results...

**Before**
|                Method |      Mean |     Error |    StdDev | Ratio | RatioSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|---------------------- |----------:|----------:|----------:|------:|--------:|--------:|--------:|--------:|----------:|
| 'System.Drawing Jpeg' |  4.498 ms |  2.060 ms | 0.1129 ms |  1.00 |    0.00 | 54.6875 | 31.2500 | 31.2500 |   2.22 MB |
|     'ImageSharp Jpeg' | 70.435 ms | 19.521 ms | 1.0700 ms | 15.66 |    0.22 |       - |       - |       - |   2.96 MB |

**After**
|                Method |      Mean |      Error |    StdDev | Ratio | RatioSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|---------------------- |----------:|-----------:|----------:|------:|--------:|--------:|--------:|--------:|----------:|
| 'System.Drawing Jpeg' |  4.495 ms |  0.6598 ms | 0.0362 ms |  1.00 |    0.00 | 54.6875 | 31.2500 | 31.2500 |   2.22 MB |
|     'ImageSharp Jpeg' | 70.219 ms | 16.7004 ms | 0.9154 ms | 15.62 |    0.19 |       - |       - |       - |   2.96 MB |

I'd take the massive difference in speed between us and System.Drawing in this benchmark with a pinch of salt. They're not subsampling which will be a huge saving.